Fix use of deallocated temporary in get_logged_users
The actual MaybeUninit here was never given a name, so it is a temporary
and thus destroyed at the semicolon. This is just UB in general, but in
particular the danger here is that the stack slot that's supposed to be
occupied by this MaybeUninit may be reused, causing all kinds of stack
corruption.